requirements.dev.txt
-r requirements.base.txt

django-debug-toolbar==3.2.4
flake8==4.0.1
black==22.3.0
coverage==6.3.2
django-coverage-plug==2.0.2
model_mommy==2.0.0